pyproject.toml
[tool.poetry]
name = "s3-certidoes"
version = "0.1.0"
description = ""
authors = ["Cleiton <chdf@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
boto3 = "^1.34.149"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
